feat: add-get-all-board - load all board feature develop

+++ src/main/java/com/project_merge/jigu_travel/board/repository/BoardJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BoardJpaRepository extends JpaRepository<Board, Integer> {
+public interface BoardJpaRepository extends JpaRepository<Board, Long> {
 
 }
